+++ stroom-app/src/test/java/stroom/test/StroomIntegrationTest.java
@@ -25,6 +25,7 @@ import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import stroom.util.io.FileUtil;
 import stroom.util.test.StroomTest;
 import stroom.util.test.TempDir;
 import stroom.util.test.TempDirExtension;
@@ -58,11 +59,6 @@ public abstract class StroomIntegrationTest implements StroomTest {
         state.reset();
     }
 
-    @BeforeEach
-    public void setup(@TempDir Path tempDir) {
-        this.testTempDir = tempDir;
-    }
-
     @AfterAll
     public static void afterClass() {
     }
@@ -82,11 +78,16 @@ public abstract class StroomIntegrationTest implements StroomTest {
      * Initialise required database entities.
      */
     @BeforeEach
-    void before(final TestInfo testInfo) {
+    void before(final TestInfo testInfo, @TempDir final Path tempDir) {
         LOGGER.info(String.format("Started test: %s::%s", testInfo.getTestClass().get().getName(), testInfo.getDisplayName()));
 
         final State state = TestState.getState();
         state.incrementTestCount();
+
+        if (tempDir == null) {
+            throw new NullPointerException("Temp dir is null");
+        }
+        this.testTempDir = tempDir;
 
         // Setup the database if this is the first test running for this test
         // class or if we always want to recreate the DB between tests.