+++ stroom-index/stroom-index-impl-db/src/test/java/stroom/index/impl/db/IndexShardDaoImplTest.java
@@ -1,4 +1,4 @@
 package stroom.index.impl.db;
 
-public class IndexShardDaoImplTest {
+class IndexShardDaoImplTest {
 }