+++ stroom-meta/stroom-meta-impl-db/src/main/resources/stroom/meta/impl/db/V7_0_0_3__meta_processor.sql
@@ -5,10 +5,11 @@
 -- between them
 CREATE TABLE IF NOT EXISTS meta_processor (
   id 				    int(11) NOT NULL AUTO_INCREMENT,
-  pipeline_uuid 	    varchar(255) NOT NULL,
-  processor_id   	    int(11) NOT NULL,
+  processor_uuid 	    varchar(255) DEFAULT NULL,
+  processor_filter_uuid varchar(255) DEFAULT NULL,
+  pipeline_uuid 	    varchar(255) DEFAULT NULL,
   PRIMARY KEY           (id),
-  UNIQUE KEY            processor_id (processor_id)
+  UNIQUE KEY            processor_uuid (processor_uuid)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 --
@@ -23,12 +24,11 @@ BEGIN
   IF (SELECT COUNT(*) FROM INFORMATION_SCHEMA.TABLES where TABLE_NAME = 'STRM_PROC' > 0) THEN
     -- Copy data into the meta_processor table, use ID predicate to make it re-runnable
     INSERT
-    INTO meta_processor (id, pipeline_uuid, processor_id)
-    SELECT SP.ID, P.UUID, SP.ID
-    FROM STRM_PROC SP
-    INNER JOIN PIPE P ON (P.ID = SP.FK_PIPE_ID)
-    WHERE SP.ID > (SELECT COALESCE(MAX(mp.id), 0) FROM meta_processor mp)
-    ORDER BY SP.ID;
+    INTO meta_processor (id, processor_uuid, pipeline_uuid)
+    SELECT ID, ID, PIPE_UUID
+    FROM STRM_PROC
+    WHERE ID > (SELECT COALESCE(MAX(id), 0) FROM meta_processor)
+    ORDER BY ID;
 
     -- Work out what to set our auto_increment start value to
     SELECT CONCAT('ALTER TABLE meta_processor AUTO_INCREMENT = ', COALESCE(MAX(id) + 1, 1))