+++ stroom-processor/stroom-processor-impl-db/src/main/java/stroom/processor/impl/db/ReprocessDataHandler.java
@@ -19,21 +19,21 @@ package stroom.processor.impl.db;
 
 import com.google.common.base.Strings;
 import stroom.entity.shared.BaseResultList;
-import stroom.meta.shared.Meta;
-import stroom.meta.shared.MetaService;
-import stroom.meta.shared.FindMetaCriteria;
-import stroom.meta.shared.MetaFieldNames;
 import stroom.entity.shared.CriteriaSet;
+import stroom.meta.shared.FindMetaCriteria;
+import stroom.meta.shared.Meta;
+import stroom.meta.shared.MetaFieldNames;
+import stroom.meta.shared.MetaService;
 import stroom.processor.StreamProcessorFilterService;
 import stroom.processor.StreamProcessorService;
+import stroom.processor.shared.Processor;
+import stroom.processor.shared.task.ReprocessDataAction;
 import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.ExpressionTerm;
 import stroom.security.Security;
 import stroom.security.shared.PermissionNames;
 import stroom.streamstore.shared.QueryData;
 import stroom.streamstore.shared.ReprocessDataInfo;
-import stroom.processor.shared.Processor;
-import stroom.streamtask.shared.ReprocessDataAction;
 import stroom.task.api.AbstractTaskHandler;
 import stroom.util.shared.Severity;
 import stroom.util.shared.SharedList;
@@ -77,9 +77,9 @@ class ReprocessDataHandler extends AbstractTaskHandler<ReprocessDataAction, Shar
                 criteria.obtainPageRequest().setOffset(0L);
                 criteria.obtainPageRequest().setLength(MAX_STREAM_TO_REPROCESS);
 
-                final BaseResultList<Meta> streams = metaService.find(criteria);
+                final BaseResultList<Meta> metaList = metaService.find(criteria);
 
-                if (!streams.isExact()) {
+                if (!metaList.isExact()) {
                     info.add(new ReprocessDataInfo(Severity.ERROR, "Results exceed " + MAX_STREAM_TO_REPROCESS
                             + " configure a pipeline processor for large data sets", null));
 
@@ -90,12 +90,11 @@ class ReprocessDataHandler extends AbstractTaskHandler<ReprocessDataAction, Shar
 
                     final Map<Processor, CriteriaSet<Long>> streamToProcessorSet = new HashMap<>();
 
-                    for (final Meta meta : streams) {
-                        // We can only reprocess streams that have a stream
-                        // processor and a parent stream id.
-                        if (meta.getProcessorId() != null && meta.getParentMetaId() != null) {
-                            final Processor streamProcessor = streamProcessorService.fetchInsecure(meta.getProcessorId());
-                            streamToProcessorSet.computeIfAbsent(streamProcessor, k -> new CriteriaSet<>()).add(meta.getParentMetaId());
+                    for (final Meta meta : metaList) {
+                        // We can only reprocess streams that have a stream processor and a parent stream id.
+                        if (meta.getProcessorUuid() != null && meta.getParentMetaId() != null) {
+                            final Processor processor = streamProcessorService.loadByUuid(meta.getProcessorUuid());
+                            streamToProcessorSet.computeIfAbsent(processor, k -> new CriteriaSet<>()).add(meta.getParentMetaId());
                         } else {
                             skippingCount++;
                         }