+++ stroom-processor/stroom-processor-impl-db/src/main/java/stroom/processor/impl/db/task/StreamProcessorTaskHandler.java
@@ -121,7 +121,7 @@ public class StreamProcessorTaskHandler extends AbstractTaskHandler<StreamProces
                     }
 
                     // Don't process any streams that we have already created
-                    if (meta.getProcessorId() != null && meta.getProcessorId() == destStreamProcessor.getId()) {
+                    if (meta.getProcessorUuid() != null && meta.getProcessorUuid().equals(destStreamProcessor.getUuid())) {
                         complete = true;
                         LOGGER.warn("Skipping data that we seem to have created (avoid processing forever) {} {}", meta,
                                 destStreamProcessor);