+++ stroom-security/stroom-security-impl-db/src/test/java/stroom/security/impl/db/AppPermissionDaoImplTest.java
@@ -19,7 +19,7 @@ import java.util.UUID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class AppPermissionDaoImplTest {
+class AppPermissionDaoImplTest {
     private static final Logger LOGGER = LoggerFactory.getLogger(AppPermissionDaoImplTest.class);
 
     private static MySQLContainer dbContainer = new MySQLContainer()
@@ -32,7 +32,7 @@ public class AppPermissionDaoImplTest {
     private static final String PERMISSION_NAME_2 = "USE_THE_FANCY_TOWELS";
 
     @BeforeAll
-    public static void beforeAll() {
+    static void beforeAll() {
         LOGGER.info(() -> "Before All - Start Database");
         Optional.ofNullable(dbContainer).ifPresent(MySQLContainer::start);
 
@@ -43,7 +43,7 @@ public class AppPermissionDaoImplTest {
     }
 
     @Test
-    public void testPermissionInvalidUser() {
+    void testPermissionInvalidUser() {
         // Given
         final String userUuid = UUID.randomUUID().toString();
 
@@ -52,7 +52,7 @@ public class AppPermissionDaoImplTest {
     }
 
     @Test
-    public void testPermissionStory() {
+    void testPermissionStory() {
         final String userName = String.format("SomePerson_%s", UUID.randomUUID());
 
         final User user = userDao.createUser(userName);
@@ -68,7 +68,7 @@ public class AppPermissionDaoImplTest {
     }
 
     @AfterAll
-    public static void afterAll() {
+    static void afterAll() {
         LOGGER.info(() -> "After All - Stop Database");
         Optional.ofNullable(dbContainer).ifPresent(MySQLContainer::stop);
     }