+++ stroom-util/src/main/java/stroom/entity/BasicCrudDao.java
@@ -1,14 +1,35 @@
 package stroom.entity;
 
+import javax.annotation.Nonnull;
 import java.util.Optional;
 
-public interface BasicCrudDao<T> {
+public interface BasicCrudDao<EntityType, IdType> {
 
-    T create();
+    /**
+     * Persist the passes record object
+     * @param record Object to persist.
+     * @return The persisted object including any changes such as auto IDs
+     */
+    EntityType create(@Nonnull final EntityType record);
 
-    T update(final T record);
+    /**
+     * Update the passed record in the database
+     * @param record The record to update.
+     * @return The record as it now appears in the database
+     */
+    EntityType update(@Nonnull final EntityType record);
 
-    int delete(final int id);
+    /**
+     * Delete the record associated with the passed id value.
+     * @param id The unique identifier for the record to delete.
+     * @return The number of records deleted.
+     */
+    int delete(final IdType id);
 
-    Optional<T> fetch(final int id);
+    /**
+     * Fetch a record from the database using its unique id value.
+     * @param id The id to uniquely identify the required record with
+     * @return The record associated with the id in the database, if it exists.
+     */
+    Optional<EntityType> fetch(final IdType id);
 }